Рефакторинг. Миграция на модульную систему Vuex.
